refactor(#147): migrate llm_service call-site imports to llm_gateway
Updates 5 import sites (4 in app code + 1 in tests) to import from
llm_gateway directly instead of going through the llm_service shim.

llm_service.py shim remains in place (#149 deletes it later); this
PR just makes call sites point at the canonical module.
