+++ backend/app/services/__init__.py
@@ -7,7 +7,7 @@ input, call a service, and format the response.
 
 from app.services.application_service import ApplicationService
 from app.services.github_service import GitHubService
-from app.services.llm_service import RewriteStrategy
+from app.services.llm_gateway import RewriteStrategy
 from app.services.pdf_service import PDFService
 from app.services.resume_parser import ResumeAST, ResumeParser
 from app.services.resume_validator import ResumeValidator, ValidationResult
